Change relationship between psychologist and user Add client to psychologist

+++ src/psychologists/psychologists.service.ts
@@ -1,17 +1,18 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Role } from 'src/roles/roles.enum';
-import { UserDocument } from 'src/users/schemas/user.schema';
-import { UsersService } from 'src/users/users.service';
 import { addItems } from 'src/common/utils/add-items.util';
 import { removeItems } from 'src/common/utils/remove-item.util';
+import { Role } from 'src/roles/roles.enum';
+import { UsersService } from 'src/users/users.service';
 import { CreatePsychologistDto } from './dto/create-psychologist.dto';
 import { UpdatePsychologistDto } from './dto/update-psychologist.dto';
 import {
   Psychologist,
   PsychologistDocument,
 } from './schemas/psychologist.schema';
+
+const submodels = ['user'];
 
 @Injectable()
 export class PsychologistsService {
@@ -21,35 +22,40 @@ export class PsychologistsService {
     private userService: UsersService,
   ) {}
 
-  async getAll(): Promise<Array<UserDocument>> {
-    return this.userService.getAllByRole(Role.Psychologist);
+  async getAll(): Promise<Array<PsychologistDocument>> {
+    return this.psychologistModel.find().populate(submodels).exec();
   }
 
-  async getById(id: string): Promise<UserDocument> {
-    return this.userService.getByPsychologistId(id);
+  async getById(id: string): Promise<PsychologistDocument | null> {
+    return this.psychologistModel.findById(id).populate(submodels).exec();
   }
 
-  async create(createData: CreatePsychologistDto): Promise<UserDocument> {
+  async create(
+    createData: CreatePsychologistDto,
+  ): Promise<PsychologistDocument> {
+    const psychologist = await this.psychologistModel
+      .findOne({ user: createData.userId })
+      .populate(submodels)
+      .exec();
+
+    if (psychologist) {
+      return psychologist;
+    }
+
     const user = await this.userService.getById(createData.userId);
 
     if (!user) {
       throw new Error('User is not found');
     }
 
-    if (user.psychologist) {
-      return user;
-    }
-
-    const psychologist = await this.psychologistModel.create({});
-    this.userService.bindWithPsychologist(
-      createData.userId,
-      psychologist._id.toString(),
-    );
+    const newPsychologist = await this.psychologistModel.create({
+      user: user._id,
+    });
 
     user.roles = addItems<Role>(user.roles, [Role.Psychologist]);
     await user.save();
 
-    return user;
+    return newPsychologist.populate(submodels);
   }
 
   async update(
@@ -57,19 +63,40 @@ export class PsychologistsService {
     updateData: UpdatePsychologistDto,
   ): Promise<PsychologistDocument> {
     await this.psychologistModel.findByIdAndUpdate(id, updateData);
-    return this.psychologistModel.findById(id);
+    return this.getById(id);
   }
 
-  async remove(id: string): Promise<PsychologistDocument> {
-    const psychologist = await this.psychologistModel.findByIdAndRemove(id);
+  async addClient(psychologistId: string, clientId: string): Promise<boolean> {
+    const psychologist = await this.getById(psychologistId);
+    const client = await this.userService.getById(clientId);
 
     if (psychologist) {
-      const user = await this.userService.getByPsychologistId(
-        psychologist._id.toString(),
+      const clients = psychologist.clients || [];
+      if (!clients.includes(client._id.toString())) {
+        clients.push(client._id);
+        psychologist.clients = clients;
+        psychologist.save();
+      }
+
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  async remove(id: string): Promise<PsychologistDocument | null> {
+    const psychologist = await this.getById(id);
+
+    if (psychologist) {
+      const user = await this.userService.getById(
+        psychologist.user._id.toString(),
       );
-      user.psychologist = undefined;
-      user.roles = removeItems<Role>(user.roles, [Role.Psychologist]);
+      user.roles = removeItems<Role>(psychologist.user.roles, [
+        Role.Psychologist,
+      ]);
       await user.save();
+
+      await this.psychologistModel.findByIdAndRemove(id);
     }
 
     return psychologist;
